Basic async tokio runtime

+++ src/feed.rs
@@ -5,5 +5,5 @@ pub struct Feed {
     pub id: u32,
     pub name: String,
     pub url: String,
-    pub frequency: u32,
+    pub frequency: u64,
 }
